Early packagist/composer layout
