+++ composer.json
@@ -1,0 +1,27 @@
+{
+    "name": "Bango29/SimpleBayesClassifier",
+    "type": "library",
+    "description": "Naive Bayes Classifier with Redis backend",
+    "keywords": ["bayes","classifier","redis"],
+    "homepage": "https://github.com/tistaharahap/Simple-Naive-Bayes-Classifier-for-PHP",
+    "license": "MIT",
+    "authors": [
+        {
+            "name": "Batista Harahap",
+            "email": "batista@bango29.com",
+            "homepage": "http://bango29.com",
+            "role": "Developer"
+        }
+    ],
+    "require": {
+        "php": ">=5.2.0"
+    },
+    "require-dev": {
+        "phpunit/phpunit":"3.7.*"
+    },
+    "autoload": {
+        "psr-0": {
+            "Bango29": "src"
+        }
+    }
+}
