+++ src/SimpleBayesClassifier/Classifier/Store/NaiveBayesClassifierStoreRedis.php
@@ -1,36 +1,39 @@
 <?php
+
+namespace \Bango29\SimpleBayesClassifier\Classifier;
+
 /**
  * Abstract implementation of NaiveBayesClassifierStore for Redis
- * 
+ *
  * @package	Simple NaiveBayesClassifier for PHP
  * @subpackage	NaiveBayesClassifierStoreRedis
  * @author	Batista R. Harahap <batista@bango29.com>
  * @link	http://www.bango29.com
  * @license	MIT License - http://www.opensource.org/licenses/mit-license.php
- * 
- * Permission is hereby granted, free of charge, to any person obtaining a 
- * copy of this software and associated documentation files (the "Software"), 
- * to deal in the Software without restriction, including without limitation 
- * the rights to use, copy, modify, merge, publish, distribute, sublicense, 
- * and/or sell copies of the Software, and to permit persons to whom the 
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a
+ * copy of this software and associated documentation files (the "Software"),
+ * to deal in the Software without restriction, including without limitation
+ * the rights to use, copy, modify, merge, publish, distribute, sublicense,
+ * and/or sell copies of the Software, and to permit persons to whom the
  * Software is furnished to do so, subject to the following conditions:
- * 
- * The above copyright notice and this permission notice shall be included in 
+ *
+ * The above copyright notice and this permission notice shall be included in
  * all copies or substantial portions of the Software.
- * 
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR 
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, 
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE 
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER 
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING 
- * FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS 
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
+ * FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS
  * IN THE SOFTWARE.
  */
 
 require_once 'NaiveBayesClassifierStore.php';
 
 class NaiveBayesClassifierStoreRedis extends NaiveBayesClassifierStore {
-	
+
 	private $conn;
 
 	private $namespace	= 'nbc-ns';
@@ -40,7 +43,7 @@ class NaiveBayesClassifierStoreRedis extends NaiveBayesClassifierStore {
 	private $cache		= "nbc-cache";
 	public $delimiter	= "_--%%--_";
 	private $wordCount	= "--count--";
-	
+
 	function __construct($conf = array()) {
 		if(empty($conf))
 			throw new NaiveBayesClassifierException(3001);
@@ -56,30 +59,30 @@ class NaiveBayesClassifierStoreRedis extends NaiveBayesClassifierStore {
 		$this->words		= "{$this->namespace}-{$this->words}";
 		$this->sets			= "{$this->namespace}-{$this->sets}";
 		$this->cache		= "{$this->namespace}-{$this->cache}";
-				
-		// Redis connection	
+
+		// Redis connection
         $this->conn = new Redis();
         $this->conn->connect($conf['db_host'], $conf['db_port']);
 		$this->conn->select(77);
 	}
-	
+
 	public function close() {
 		$this->conn->close();
 	}
-	
+
 	public function addToBlacklist($word) {
 		return $this->conn->incr("{$this->blacklist}#{$word}");
 	}
-	
+
 	public function removeFromBlacklist($word) {
 		return $this->conn->set("{$this->blacklist}#{$word}", 0);
 	}
-	
+
 	public function isBlacklisted($word) {
 		$res = $this->conn->get("{$this->blacklist}#{$word}");
 		return !empty($res) && $res > 0 ? TRUE : FALSE;
 	}
-	
+
 	public function trainTo($word, $set) {
 		// Words
 		$this->conn->hIncrBy($this->words, $word, 1);
@@ -114,27 +117,27 @@ class NaiveBayesClassifierStoreRedis extends NaiveBayesClassifierStore {
 			return FALSE;
 		}
 	}
-	
+
 	public function getAllSets() {
 		return $this->conn->hKeys($this->sets);
 	}
-	
+
 	public function getSetCount() {
 		return $this->conn->hLen($this->sets);
 	}
-	
+
 	public function getWordCount($words) {
 		return $this->conn->hMGet($this->words, $words);
 	}
-	
+
 	public function getAllWordsCount() {
 		return $this->conn->hGet($this->wordCount, $this->wordCount);
 	}
-	
+
 	public function getSetWordCount($sets) {
 		return $this->conn->hMGet($this->sets, $sets);
 	}
-	
+
 	public function getWordCountFromSet($words, $sets) {
 		$keys = array();
 		foreach($words as $word) {
@@ -144,5 +147,5 @@ class NaiveBayesClassifierStoreRedis extends NaiveBayesClassifierStore {
 		}
 		return $this->conn->hMGet($this->words, $keys);
 	}
-	
+
 }
